refactor : user 도메인의 애플리케이션 계층 디렉토리 구조 리팩토링
