+++ src/main/java/com/pintoss/auth/core/user/application/OAuth2UseCase.java
@@ -10,9 +10,9 @@ import com.pintoss.auth.api.auth.dto.OAuth2SignupRequired;
 import com.pintoss.auth.core.user.domain.LoginType;
 import com.pintoss.auth.core.user.domain.User;
 import com.pintoss.auth.core.user.application.dto.OAuth2UserInfoResponse;
-import com.pintoss.auth.core.user.core.OAuth2UserInfoService;
-import com.pintoss.auth.core.user.core.OAuth2UserInfoStrategy;
-import com.pintoss.auth.core.user.core.UserReader;
+import com.pintoss.auth.core.user.application.flow.policy.OAuth2UserInfoService;
+import com.pintoss.auth.core.user.application.flow.policy.OAuth2UserInfoStrategy;
+import com.pintoss.auth.core.user.application.flow.reader.UserReader;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
 import lombok.RequiredArgsConstructor;
