+++ src/main/java/com/pintoss/auth/core/user/application/PasswordResetUseCase.java
@@ -3,8 +3,8 @@ package com.pintoss.auth.core.user.application;
 import com.pintoss.auth.core.user.domain.LoginType;
 import com.pintoss.auth.core.user.domain.Phone;
 import com.pintoss.auth.core.user.domain.User;
-import com.pintoss.auth.core.user.core.MailSender;
-import com.pintoss.auth.core.user.core.UserReader;
+import com.pintoss.auth.core.user.application.flow.external.MailSender;
+import com.pintoss.auth.core.user.application.flow.reader.UserReader;
 import java.security.SecureRandom;
 import java.util.Collections;
 import java.util.List;
