+++ src/main/java/com/pintoss/auth/core/user/application/UserValidationUseCase.java
@@ -1,6 +1,6 @@
 package com.pintoss.auth.core.user.application;
 
-import com.pintoss.auth.core.user.core.UserValidator;
+import com.pintoss.auth.core.user.application.flow.validator.UserValidator;
 import com.pintoss.auth.core.user.domain.Phone;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
