+++ src/main/java/com/pintoss/auth/core/user/application/flow/external/NiceAuthRequestClient.java
@@ -1,4 +1,4 @@
-package com.pintoss.auth.core.user.core;
+package com.pintoss.auth.core.user.application.flow.external;
 
 import com.pintoss.auth.api.auth.dto.IdentityVerifyPurpose;
 import com.pintoss.auth.core.user.application.dto.NiceEncryptedDataResult;
