+++ src/main/java/com/pintoss/auth/core/user/application/flow/policy/OAuth2UserInfoService.java
@@ -1,4 +1,4 @@
-package com.pintoss.auth.core.user.core;
+package com.pintoss.auth.core.user.application.flow.policy;
 
 import com.pintoss.auth.core.user.application.dto.OAuth2UserInfoResponse;
