+++ src/main/java/com/pintoss/auth/core/user/application/flow/policy/OAuth2UserInfoStrategy.java
@@ -1,4 +1,4 @@
-package com.pintoss.auth.core.user.core;
+package com.pintoss.auth.core.user.application.flow.policy;
 
 import com.pintoss.auth.common.exception.ErrorCode;
 import com.pintoss.auth.common.exception.client.BadRequestException;
