+++ src/main/java/com/pintoss/auth/core/user/application/flow/reader/UserReader.java
@@ -1,7 +1,8 @@
-package com.pintoss.auth.core.user.core;
+package com.pintoss.auth.core.user.application.flow.reader;
 
 import com.pintoss.auth.common.exception.ErrorCode;
 import com.pintoss.auth.common.exception.client.BadRequestException;
+import com.pintoss.auth.core.user.application.repository.UserRepository;
 import com.pintoss.auth.core.user.domain.LoginType;
 import com.pintoss.auth.core.user.domain.Phone;
 import com.pintoss.auth.core.user.domain.User;
