+++ src/main/java/com/pintoss/auth/core/user/application/flow/validator/UserValidator.java
@@ -1,8 +1,9 @@
-package com.pintoss.auth.core.user.core;
+package com.pintoss.auth.core.user.application.flow.validator;
 
 import com.pintoss.auth.common.exception.ErrorCode;
 import com.pintoss.auth.common.exception.client.BadRequestException;
 import com.pintoss.auth.api.security.SecurityContextUtils;
+import com.pintoss.auth.core.user.application.repository.UserRepository;
 import com.pintoss.auth.core.user.domain.Phone;
 import com.pintoss.auth.core.user.domain.User;
 import lombok.RequiredArgsConstructor;
